Fixed ship classes and properly inherit from IShip
All ships extend IShip and need to work on checkDamage, as well implemented Serializable into IShip
